# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.AddItemRequest:
    properties:
      product_id:
        type: string
      quantity:
        minimum: 1
        type: integer
      unit_price:
        minimum: 0
        type: number
    required:
    - product_id
    - quantity
    - unit_price
    type: object
...
  models.Cart:
    properties:
      created_at:
        type: string
      id:
        type: string
      items:
        additionalProperties:
          $ref: '#/definitions/models.CartItem'
        type: object
      total:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.CartItem:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
      total_price:
        type: number
      unit_price:
        type: number
    type: object
...
  models.EmailNotificationRequest:
    properties:
      bcc:
        items:
          type: string
        type: array
      cc:
        items:
          type: string
        type: array
      content:
        type: string
      html_content:
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      subject:
        type: string
      to:
        type: string
    required:
    - content
    - subject
    - to
    type: object
...
  models.Notification:
    properties:
      content:
        type: string
      created_at:
        type: string
      error_message:
        type: string
      id:
        type: string
      metadata:
        description: highly dynamic
        type: object
      recipient:
        type: string
      sent_at:
        type: string
      status:
        $ref: '#/definitions/models.NotificationStatus'
      subject:
        type: string
      type:
        $ref: '#/definitions/models.NotificationType'
      updated_at:
        type: string
    type: object
  models.NotificationStatus:
    enum:
    - pending
    - sent
    - failed
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusSent
    - StatusFailed
  models.NotificationType:
    enum:
    - email
    - sms
    - push
    type: string
    x-enum-varnames:
    - NotificationTypeEmail
    - NotificationTypeSMS
    - NotificationTypePush
...
  models.PaginatedResponse:
    properties:
      data: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  models.UpdateQuantityRequest:
    properties:
      product_id:
        type: string
      quantity:
        minimum: 0
        type: integer
    required:
    - product_id
    - quantity
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        type: string
      details:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
host: localhost:8085
info:
  contact:
    email: aaravmahajan2003@gmail.com
    name: Aarav Mahajan
    url: https://github.com/aaravmahajanofficial
  description: This is the API server for the Scalable E-commerce Platform. It provides
    endpoints for managing users, products, carts, orders, payments, and notifications.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Scalable E-commerce Platform API
  version: "1.0"
paths:
  /carts:
    get:
      description: Retrieves the current shopping cart contents for the authenticated
        user. Creates a cart if one doesn't exist.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved or created cart
          schema:
            $ref: '#/definitions/models.Cart'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error (e.g., failed to create cart)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get the user's shopping cart
      tags:
      - Cart
  /carts/items:
    post:
      consumes:
      - application/json
      description: Adds a specified quantity of a product to the authenticated user's
        shopping cart. Creates cart if needed.
      parameters:
      - description: Item details (Product ID and Quantity)
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.AddItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Item successfully added/updated in cart
          schema:
            $ref: '#/definitions/models.Cart'
        "400":
          description: Validation error or invalid product ID/quantity
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add an item to the cart
      tags:
      - Cart
    put:
      consumes:
      - application/json
      description: Updates the quantity of a specific item in the authenticated user's
        shopping cart.
      parameters:
      - description: Item details (Product ID and new Quantity)
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.UpdateQuantityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Quantity successfully updated
          schema:
            $ref: '#/definitions/models.Cart'
        "400":
          description: Validation error or invalid product ID/quantity
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Cart or item not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update item quantity in the cart
      tags:
      - Cart
  /notifications:
    get:
      description: Retrieves a paginated list of notifications for the authenticated
        user. Requires authentication.
      parameters:
      - description: 'Page number for pagination (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 'Number of items per page (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved list of notifications
          schema:
            allOf:
            - $ref: '#/definitions/models.PaginatedResponse'
            - properties:
                Data:
                  items:
                    $ref: '#/definitions/models.Notification'
                  type: array
              type: object
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List notifications for the user
      tags:
      - Notifications
  /notifications/email:
    post:
      consumes:
      - application/json
      description: Creates and sends an email notification record. This might be an
        admin-triggered action or for specific internal purposes. Requires authentication.
      parameters:
      - description: Email Notification Details (Recipient User ID, Subject, Body)
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/models.EmailNotificationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created and potentially queued email notification
          schema:
            $ref: '#/definitions/models.Notification'
        "400":
          description: Validation error or invalid input
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden - Insufficient permissions" // If restricted
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Recipient User not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error or email sending provider error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Send an email notification (Admin/Internal)
      tags:
      - Notifications
...
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"